refactor: 코드리펙토링
- Operator변경에 따른 영향도 줄이기.
- 연산자에 대한 Map을 static 필드로 관리하도록..
- Validation
+ Filter의 기능을 Parser로 합침.
- 입, 출력 분리
