+++ src/main/java/calculator/engine/io/Input.java
@@ -1,0 +1,5 @@
+package calculator.engine.io;
+
+public interface Input {
+    String input(String prompt);
+}
