+++ src/main/java/calculator/service/Parser.java
@@ -1,0 +1,5 @@
+package calculator.service;
+
+public interface Parser {
+    String parse(String command);
+}
