+++ src/main/java/calculator/service/Sorter.java
@@ -1,0 +1,9 @@
+package calculator.service;
+
+import calculator.engine.model.OperatorOrder;
+
+import java.util.List;
+
+public interface Sorter {
+    List<OperatorOrder> sort(List<OperatorOrder> list);
+}
