--- src/main/java/repository/CalculationRepository.java
@@ -1,10 +1,0 @@
-package repository;
-
-import model.CalculationDto;
-
-import java.util.List;
-
-public interface CalculationRepository {
-    void save(CalculationDto calculationDto);
-    List<String> findAll();
-}
