--- src/main/java/service/Calculate.java
@@ -1,5 +1,0 @@
-package service;
-
-public interface Calculate {
-    double calc(String command);
-}
