--- src/main/java/service/CalculationService.java
@@ -1,10 +1,0 @@
-package service;
-
-import model.CalculationDto;
-
-import java.util.List;
-
-public interface CalculationService {
-    List<String> findAll();
-    CalculationDto calculate(String command);
-}
