--- src/main/java/service/CommandFilter.java
@@ -1,5 +1,0 @@
-package service;
-
-public interface CommandFilter {
-    String filter(String command);
-}
