--- src/main/java/service/Sorter.java
@@ -1,9 +1,0 @@
-package service;
-
-import model.OperatorOrder;
-
-import java.util.List;
-
-public interface Sorter {
-    List<OperatorOrder> sort(List<OperatorOrder> list);
-}
