--- src/main/java/service/ValidationCheck.java
@@ -1,5 +1,0 @@
-package service;
-
-public interface ValidationCheck {
-    boolean validate(String command);
-}
